+++ chaos/Image/include/Def.h
@@ -1,0 +1,15 @@
+//Def.h
+#ifndef CHAOS_IMAGE_INCLUDE_DEF_H
+#define CHASO_IAMGE_INCLUDE_DEF_H
+
+namespace chaos {
+
+namespace image {
+
+typedef unsigned char uint8_t;
+
+}  //namespace image
+
+}  //namespace chaos
+
+#endif