+++ Changelog.md
@@ -1,4 +1,5 @@
 # chaos changelog
+
 ### 2022.02.21
 * add Arithmetic lib;
 
@@ -15,3 +16,11 @@
 
 ### 2022.03.06
 * add: Dynamic memory trigger, if size less than 16, heap memory would be used.
+
+### 2022.03.11
+* add: Image lib. Image class and load && save function.
+
+### 2022.03.12
+* add: Geometry lib. Base type such as Point2 and Line2;
+* add: Drawer for Image, implement plot line function;
+* add: Test for Drawer;